Separate the article action menu
